# Optimize stack usage measuring