--- a/src/canary.rs
+++ b/src/canary.rs
@@ -115,21 +115,19 @@ impl Canary {
         if self.measure_stack {
             log::info!("reading {size_kb:.2} KiB of RAM for stack usage estimation");
         }
-        let mut canary = vec![0; self.size];
         let start = Instant::now();
-        core.read_8(self.address.into(), &mut canary)?;
+        let touched_address =
+            measure_subroutine::execute(core, self.address as u64, self.size as u64)?;
         let seconds = start.elapsed().as_secs_f64();
         log::trace!(
             "reading canary took {seconds:.3}s ({:.2} KiB/s)",
             size_kb / seconds
         );
 
-        let min_stack_usage = match canary.iter().position(|b| *b != CANARY_U8) {
-            Some(pos) => {
-                let touched_address = self.address + pos as u32;
+        let min_stack_usage = match touched_address {
+            Some(touched_address) => {
                 log::debug!("canary was touched at {touched_address:#010X}");
-
-                Some(elf.vector_table.initial_stack_pointer - touched_address)
+                Some(elf.vector_table.initial_stack_pointer - touched_address as u32)
             }
             None => None,
         };
@@ -263,6 +261,120 @@ mod paint_subroutine {
         0xfb, 0xe7, // b.n      200ec <paint>
         0x00, 0xbe, // bkpt     0x0000
         0x00, 0xbe, // bkpt     0x0000 (padding instruction)
+    ];
+}
+
+/// Search for lowest touched address in memory.
+///
+/// ### Corresponds to following rust code
+///
+/// ```rust
+/// #[export_name = "measure"]
+/// unsafe fn measure(mut low_addr: u32, high_addr: u32, pattern: u32) -> u32 {
+///     let mut result = 0;
+///
+///     loop {
+///         if low_addr >= high_addr {
+///             break;
+///         } else if (low_addr as *const u32).read() != pattern {
+///             result = low_addr;
+///             break;
+///         } else {
+///             low_addr += 4;
+///             continue;
+///         }
+///     }
+///
+///     result
+/// }
+/// ```
+///
+/// ### Generated assembly
+///
+/// The assembly is generated from abothes rust code, using the jorge-hack. The labels
+/// `<end>` and `<no_result>` got manually inserted for readability.
+///
+/// ```armasm
+/// 000200ec <measure>:
+///     200ec:    4288    cmp      r0, r1
+///     200ee:    d204    bcs.n    200fa <no_result>
+///     200f0:    6803    ldr      r3, [r0, #0]
+///     200f2:    4293    cmp      r3, r2
+///     200f4:    d102    bne.n    200fc <end>
+///     200f6:    1d00    adds     r0, r0, #4
+///     200f8:    e7f8    b.n      200ec <measure>
+/// 000200fa <no_result>:
+///     200fa:    2000    movs     r0, #0
+/// 000200fc <end>:
+///     200fc:    be00    bkpt     0x0000
+/// //                    ^^^^ this was `bx lr`
+/// ```
+///
+/// ### Register-parameter-mapping
+///
+/// - r0: low_addr, and return value
+/// - r1: high_addr
+/// - r2: pattern
+mod measure_subroutine {
+    use super::*;
+
+    /// Execute the subroutine.
+    ///
+    /// The returned `Option<u32>` is None, if the memory is untouched. Otherwise it
+    /// gives the position of the first (lowest) 4-byte-word which isn't the same as
+    /// the pattern anymore. Because we are searching for 4-byte-words, the returned
+    /// `u32` will always be a multiple of 4.
+    ///
+    /// ## Assumptions
+    /// - Expects the [`Core`] to be halted and will leave it halted when the function
+    /// returns.
+    /// - `low_addr` and `size` need to be 4-byte-aligned.
+    pub fn execute(
+        core: &mut Core,
+        low_addr: u64,
+        stack_size: u64,
+    ) -> Result<Option<u64>, probe_rs::Error> {
+        assert_subroutine!(low_addr, stack_size, self::SUBROUTINE.len() as u64);
+
+        // use probe to search through the memory the subroutine will be written to
+        // NOTE: SUBROUTINE.len() is always a multiple of 4, therefore we can divide by 4 here
+        let mut buf = [0; self::SUBROUTINE.len()];
+        core.read_8(low_addr, &mut buf)?;
+        match buf.iter().position(|b| *b != CANARY_U8) {
+            // if we find a touched value, we return early
+            Some(pos) => return Ok(Some(pos as u64)),
+            // otherwise continue
+            None => {}
+        }
+
+        // prepare subroutine
+        let _ = super::prepare_subroutine(core, low_addr, stack_size, self::SUBROUTINE)?;
+
+        // execute the subroutine and wait for it to finish
+        core.run()?;
+        core.wait_for_core_halted(TIMEOUT)?;
+
+        // read out and return the result
+        match core.read_core_reg(RegisterId(0))? {
+            0 => Ok(None),
+            n => Ok(Some(n - low_addr)),
+        }
+    }
+
+    const SUBROUTINE: [u8; 20] = [
+        // <measure>
+        0x88, 0x42, // cmp     r0, r1
+        0x04, 0xd2, // bcs.n   200fa <no_result>
+        0x03, 0x68, // ldr     r3, [r0, #0]
+        0x93, 0x42, // cmp     r3, r2
+        0x02, 0xd1, // bne.n   200fc <end>
+        0x00, 0x1d, // adds    r0, r0, #4
+        0xf8, 0xe7, // b.n     200ec <measure>
+        // <no_result>
+        0x00, 0x20, // movs    r0, #0
+        // <end>
+        0x00, 0xbe, // bkpt    0x0000
+        0x00, 0xbe, // bkpt    0x0000 (padding instruction)
     ];
 }