--- a/src/canary.rs
+++ b/src/canary.rs
@@ -341,7 +341,7 @@ mod measure_subroutine {
         let mut buf = [0; self::SUBROUTINE.len()];
         core.read_8(low_addr as u64, &mut buf)?;
         match buf.iter().position(|b| *b != CANARY_U8) {
-            Some(pos) => Ok(Some(pos as u32)),
+            Some(pos) => Ok(Some(low_addr + pos as u32)),
             None => Ok(None),
         }
     }