--- a/src/canary.rs
+++ b/src/canary.rs
@@ -218,7 +218,7 @@ macro_rules! assert_subroutine {
 ///    200ec:    4288    cmp      r0, r1
 ///    200ee:    d801    bhi.n    #6 <paint+0x8>
 ///    200f0:    c004    stmia    r0!, {r2}
-///    200f2:    e7fb    b.n      #0xfffffffa <paint>
+///    200f2:    e7fb    b.n      #-6 <paint>
 ///
 /// 000200f4 <paint+0x8>:
 ///    200f4:    be00    bkpt     0x0000
@@ -253,7 +253,7 @@ mod paint_subroutine {
         0x88, 0x42, // cmp      r0, r1
         0x01, 0xd8, // bhi.n    #6 <paint+0x8>
         0x04, 0xc0, // stmia    r0!, {r2}
-        0xfb, 0xe7, // b.n      #0xfffffffa <paint>
+        0xfb, 0xe7, // b.n      #-6 <paint>
         0x00, 0xbe, // bkpt     0x0000
         0x00, 0xbe, // bkpt     0x0000 (padding instruction)
     ];
@@ -293,7 +293,7 @@ mod paint_subroutine {
 ///     200f2:    4293    cmp      r3, r2
 ///     200f4:    d102    bne.n    #8 <measure+0x10>
 ///     200f6:    1d00    adds     r0, r0, #4
-///     200f8:    e7f8    b.n      #0xfffffff8 <measure>
+///     200f8:    e7f8    b.n      #-8 <measure>
 ///
 /// 000200fa <measure+0xe>:
 ///     200fa:    2000    movs     r0, #0
@@ -371,7 +371,7 @@ mod measure_subroutine {
         0x93, 0x42, // cmp      r3, r2
         0x02, 0xd1, // bne.n    #8 <measure+0x10>
         0x00, 0x1d, // adds     r0, r0, #4
-        0xf8, 0xe7, // b.n      #0xfffffff8 <measure>
+        0xf8, 0xe7, // b.n      #-8 <measure>
         0x00, 0x20, // movs     r0, #0
         0x00, 0xbe, // bkpt     0x0000
         0x00, 0xbe, // bkpt     0x0000 (padding instruction)